Compilable program, readz to be tested

+++ SW/custom_inst.h
@@ -3,6 +3,6 @@
 #ifndef CUSTOM_INSTRUCTIONS_H
 #define CUSTOM_INSTRUCTIONS_H
 
-uint32_t mul16(uint16_t a, uint16_t b);
+uint32_t mul16(uint32_t a, uint32_t b);
 
 #endif
